moved "Dataset" into core
