+++ hub/api/tensor.py
@@ -1,10 +1,9 @@
-from hub.util.keys import get_chunk_id_encoder_key, get_tensor_meta_key, tensor_exists
+from hub.util.keys import tensor_exists
 from hub.core.sample import Sample  # type: ignore
-from typing import List, Sequence, Union, Optional, Tuple, Dict
+from typing import List, Sequence, Union, Optional, Tuple
 from hub.util.shape import ShapeInterval
 
 import numpy as np
-
 from hub.core.chunk_engine import ChunkEngine, SampleValue
 from hub.core.storage import LRUCache
 from hub.util.exceptions import (
