+++ hub/constants.py
@@ -29,7 +29,7 @@ DEFAULT_MAX_CHUNK_SIZE = 32 * MB
 MIN_FIRST_CACHE_SIZE = 32 * MB
 MIN_SECOND_CACHE_SIZE = 160 * MB
 
-# without MB multiplication, meant for the Dataset API that takes cache size in MBs
+# without MB multiplication, meant for the dataset API that takes cache size in MBs
 DEFAULT_MEMORY_CACHE_SIZE = 256
 DEFAULT_LOCAL_CACHE_SIZE = 0
