+++ hub/util/from_tfds.py
@@ -5,7 +5,7 @@ import tensorflow_datasets as tfds  # type: ignore
 from tensorflow_datasets import Split
 from tqdm import tqdm  # type: ignore
 
-from hub.api.dataset import Dataset
+from hub.core.dataset import Dataset
 import hub
